+++ platform/windows/capi/include/ag_dns.h
@@ -244,6 +244,8 @@ typedef struct {
      * the proxy will respond with a SERVFAIL packet. Otherwise, no response is sent on such a failure.
      */
     bool enable_servfail_on_upstreams_failure;
+    /** Enable HTTP/3 for DNS-over-HTTPS upstreams if it's able to connect quicker. */
+    bool enable_http3;
 } ag_dnsproxy_settings;
 
 typedef struct {