+++ platform/windows/capi/src/ag_dns.cpp
@@ -275,6 +275,7 @@ static ag_dnsproxy_settings *marshal_settings(const DnsProxySettings &settings) 
     c_settings->enable_parallel_upstream_queries = settings.enable_parallel_upstream_queries;
     c_settings->enable_fallback_on_upstreams_failure = settings.enable_fallback_on_upstreams_failure;
     c_settings->enable_servfail_on_upstreams_failure = settings.enable_servfail_on_upstreams_failure;
+    c_settings->enable_http3 = settings.enable_http3;
 
     return c_settings;
 }
@@ -462,6 +463,7 @@ static DnsProxySettings marshal_settings(const ag_dnsproxy_settings *c_settings)
     settings.enable_parallel_upstream_queries = c_settings->enable_parallel_upstream_queries;
     settings.enable_fallback_on_upstreams_failure = c_settings->enable_fallback_on_upstreams_failure;
     settings.enable_servfail_on_upstreams_failure = c_settings->enable_servfail_on_upstreams_failure;
+    settings.enable_http3 = c_settings->enable_http3;
 
     return settings;
 }